xilffs: Fix stack memory access issues
Due to the incorrect memory access while reading the Sector Size
information, leads to memory corruption of the other local stack
variables(sz_blk) in f_mkfs() function. This results to file system
errors while formatting the disk.
Update the diskio.c, ff.c files to fix the memory access issues.


Acked-for-series: Srinivas Goud <srinivas.goud@amd.com>

+++ lib/sw_services/xilffs/src/diskio.c
@@ -599,7 +599,7 @@ DRESULT disk_ioctl (
 		case (BYTE)GET_SECTOR_SIZE : /* Get sector size on the disk (DWORD) */
 			if (pdrv < XSDPS_NUM_INSTANCES) {
 #ifdef XPAR_XSDPS_NUM_INSTANCES
-				(*((DWORD*)((void *)buff))) = ((DWORD)512U);
+				(*((WORD*)((void *)buff))) = ((WORD)512U);
 #endif
 			} else {
 #ifdef XPAR_XUFSPSXC_NUM_INSTANCES
@@ -615,7 +615,7 @@ DRESULT disk_ioctl (
 					LunId = pdrv - XUFSPSXC_START_INDEX;
 				}
 
-				(*((DWORD*)((void *)buff))) = ((DWORD)UfsInstance.LUNInfo[LunId].BlockSize);
+				(*((WORD*)((void *)buff))) = ((WORD)UfsInstance.LUNInfo[LunId].BlockSize);
 #endif
 			}
 			res = RES_OK;
